+++ fury/ui/__init__.py
@@ -1,4 +1,4 @@
 
 from fury.ui.core import *
-from fury.ui.components import *
 from fury.ui.containers import *
+from fury.ui.elements import *
